[MOV] im_livechat: move WebsiteLivechatWindow outside of -
- public_livechat.js

Task-2212347
